+++ frontend/src/modules/InplaceVolumetrics/view/view.tsx
@@ -1,24 +1,18 @@
 import React from "react";
-import Plot from "react-plotly.js";
 
-import { InplaceVolumetricData_api, InplaceVolumetricResponseNames_api } from "@api";
-import { EnsembleIdent } from "@framework/EnsembleIdent";
-import { ParameterIdent } from "@framework/EnsembleParameters";
+import { InplaceVolumetricResponseNames_api } from "@api";
 import { EnsembleSet } from "@framework/EnsembleSet";
 import { ModuleViewProps } from "@framework/Module";
-import { EnsembleRealizationFilterFunction, useEnsembleRealizationFilterFunc } from "@framework/WorkbenchSession";
+import { useEnsembleRealizationFilterFunc } from "@framework/WorkbenchSession";
 import { useElementSize } from "@lib/hooks/useElementSize";
 import { ColorSet } from "@lib/utils/ColorSet";
-import { makeSubplots } from "@modules/_shared/Figure";
-import { computeQuantile } from "@modules/_shared/statistics";
 
-import { group } from "console";
-
-import { InplaceHistogramPlot, InplaceResultValues, addHistogramTrace } from "./components/inplaceHistogramPlot";
+import { InplaceDistributionPlot, InplaceResultValues } from "./components/InplaceDistributionPlot";
 import { useInplaceDataResultsQuery } from "./hooks/queryHooks";
 
 import { Interface } from "../settingsToViewInterface";
 import { State } from "../state";
+import { PlotTypeEnum } from "../typesAndEnums";
 import { InplaceVolGroupedResultValues, getGroupedInplaceVolResults } from "../utils/inplaceVolDataEnsembleSetAccessor";
 
 export function View(props: ModuleViewProps<State, Interface>) {
@@ -43,7 +37,7 @@ export function View(props: ModuleViewProps<State, Interface>) {
         selectedInplaceResponseName as InplaceVolumetricResponseNames_api,
         selectedInplaceIndexesValues
     );
-
+    const plotType = props.viewContext.useSettingsToViewInterfaceValue("plotType");
     const colorBy = props.viewContext.useSettingsToViewInterfaceValue("colorBy");
     const groupBy = props.viewContext.useSettingsToViewInterfaceValue("groupBy");
 
@@ -63,7 +57,8 @@ export function View(props: ModuleViewProps<State, Interface>) {
     };
     return (
         <div className="w-full h-full" ref={wrapperDivRef}>
-            <InplaceHistogramPlot
+            <InplaceDistributionPlot
+                plotType={plotType}
                 resultValues={resultValues}
                 width={wrapperDivSize.width}
                 height={wrapperDivSize.height}
